django_project/settings.py
from pathlib import Path

from decouple import Csv, config
from django.core.management.utils import get_random_secret_key
from django.utils.translation import gettext_lazy as _

BASE_DIR = Path(__file__).resolve().parent.parent
SECRET_KEY = config("SECRET_KEY", default=get_random_secret_key(), cast=str)
DEBUG = config("DEBUG", default=False, cast=bool)
ALLOWED_HOSTS = config("ALLOWED_HOSTS", cast=Csv())

DATABASES = {
    "default": {
        "ENGINE": config("DB_ENGINE", cast=str),
        "NAME": BASE_DIR / config("DB_NAME", cast=str),
    },
    "postgresql": {
        "ENGINE": config("DB_ENGINE", cast=str),
        "NAME": config("DB_NAME", cast=str),
        "USER": config("DB_USER", cast=str),
        "PASSWORD": config("DB_PASSWORD", cast=str),
        "HOST": config("DB_HOST", cast=str),
        "PORT": config("DB_PORT", cast=int),
    }
}

LANGUAGE_CODE = "en"

LANGUAGES = [
    ("uz", _("O'zbekcha")),
    ("ru", _("Русский")),
    ("en", _("English")),
]

LOCALE_PATHS = [
    BASE_DIR / "app_1" / "locale",
    BASE_DIR / "app_2" / "locale",
    BASE_DIR / "app_n" / "locale",
]

TIME_ZONE = "Moscow/Russia"
USE_I18N = True
USE_L10N = True
USE_TZ = True

STATIC_URL = "static/"
STATIC_ROOT = BASE_DIR / "staticfiles/"
STATICFILES_DIRS = [BASE_DIR / "static/"]

MEDIA_URL = "media/"
MEDIA_ROOT = BASE_DIR / "media/"
